requirements.txt
llama_cpp_python
python-dotenv
pandas